requirements.txt
python-dotenv~=1.2.1
httpx~=0.28.1
aiogram~=3.25.0
aiogram-dialog~=2.5.0
asyncpg~=0.31.0
pydantic~=2.12.5
requests~=2.32.5
fastapi~=0.135.1